strawberry-gql/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#

anyio==3.5.0
    # via
    #   httpcore
    #   watchgod
asgiref==3.5.0
    # via uvicorn
backports-cached-property==1.0.1
    # via strawberry-graphql
certifi==2021.10.8
    # via
    #   httpcore
    #   httpx
charset-normalizer==2.0.12
    # via httpx
click==8.1.3
    # via
    #   httpx
    #   strawberry-graphql
    #   uvicorn
colorama==0.4.4
    # via rich
commonmark==0.9.1
    # via rich
edgedb==0.23.0
    # via -r requirements.in
graphql-core==3.2.1
    # via strawberry-graphql
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.14.7
    # via httpx
httptools==0.4.0
    # via uvicorn
httpx[cli]==0.22.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   rfc3986
pygments==2.12.0
    # via
    #   httpx
    #   rich
    #   strawberry-graphql
python-dateutil==2.8.2
    # via strawberry-graphql
python-dotenv==0.20.0
    # via uvicorn
python-multipart==0.0.5
    # via strawberry-graphql
pyyaml==6.0
    # via uvicorn
rfc3986[idna2008]==1.5.0
    # via httpx
rich==10.16.2
    # via httpx
six==1.16.0
    # via
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
strawberry-graphql==0.109.1
    # via -r requirements.in
typing-extensions==4.2.0
    # via
    #   edgedb
    #   strawberry-graphql
uvicorn[standard]==0.17.6
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
watchgod==0.8.2
    # via uvicorn
websockets==10.3
    # via uvicorn
